config/dev-requirements.txt
docutils
ipython
ipdb
django-debug-toolbar